+++ apps/who-is-miles-guo/messages/cn.json
@@ -1,8 +1,9 @@
 {
   "HomePage": {
-    "title": "谁是郭文贵?"
+    "title": "谁是郭文贵?",
+    "readMore": "了解更多"
   },
-  "Metadata":{
+  "Metadata": {
     "description": "郭文贵是中国著名的商人和政治活动家，他以揭露中国政府腐败和人权侵犯而闻名。"
   }
 }
