+++ apps/who-is-miles-guo/messages/en.json
@@ -1,6 +1,7 @@
 {
   "HomePage": {
-    "title": "Who is Miles Guo?"
+    "title": "Who is Miles Guo?",
+    "readMore": "Read more"
   },
   "Metadata": {
     "description": "Miles Guo is a prominent Chinese businessman and political activist known for exposing corruption and human rights abuses within the Chinese government."
