+++ apps/who-is-miles-guo/src/app/layout.tsx
@@ -1,14 +1,10 @@
 import '@/app/globals.css';
 import 'easymde/dist/easymde.min.css';
 
-export default async function StudioLayout({
+export default async function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  return (
-    <html lang="en">
-      <body>{children}</body>
-    </html>
-  );
+  return children;
 }
