+++ apps/who-is-miles-guo/src/components/banner.tsx
@@ -7,9 +7,7 @@ export default function Banner() {
   return (
     <div className="mt-10 flex flex-col items-center justify-center gap-5 text-center">
       <h3 className="font-semibold">WiMG</h3>
-      <h1>{t('title')}</h1>
-
-      <h2 className="text-5xl font-bold tracking-wide">Who is Miles Guo?</h2>
+      <h2 className="text-5xl font-bold tracking-wide">{t('title')}</h2>
       <p className="tracking-wide">
         Read along to understand Miles Guo better, so to understand the word.
       </p>
